requirements.txt
asgiref==3.7.2
Django==5.0.1
django-autoslug==1.9.9
django-cities-light==3.10.0
django-countries==7.5.1
djangorestframework==3.14.0
progressbar2==4.3.2
python-utils==3.8.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
Unidecode==1.3.7
